Move GCP functions to dedicated lib dir
